+++ build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 allprojects {
-    group = 'site.udtk.engine'
+    group = 'com.underthekey.engine'
     version = '0.0.1-SNAPSHOT'
 
     repositories {
