+++ config-server/src/main/java/com/underthekey/configserver/ConfigServerApplication.java
@@ -1,0 +1,19 @@
+package com.underthekey.configserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.config.server.EnableConfigServer;
+
+@SpringBootApplication(
+	scanBasePackages = "site.udtk.globalsecurity"
+)
+@EnableConfigServer
+@EnableDiscoveryClient
+public class ConfigServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ConfigServerApplication.class, args);
+	}
+
+}
