--- config-server/src/main/java/site/udtk/configserver/ConfigServerApplication.java
@@ -1,19 +1,0 @@
-package site.udtk.configserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.config.server.EnableConfigServer;
-
-@SpringBootApplication(
-	scanBasePackages = "site.udtk.globalsecurity"
-)
-@EnableConfigServer
-@EnableDiscoveryClient
-public class ConfigServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ConfigServerApplication.class, args);
-	}
-
-}
