+++ config-server/src/test/java/com/underthekey/configserver/ConfigServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.underthekey.configserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ConfigServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
