--- config-server/src/test/java/site/udtk/configserver/ConfigServerApplicationTests.java
@@ -1,13 +1,0 @@
-package site.udtk.configserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ConfigServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
