+++ core/src/main/java/com/underthekey/core/CoreApplication.java
@@ -1,0 +1,15 @@
+package com.underthekey.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class CoreApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CoreApplication.class, args);
+	}
+
+}
