--- core/src/main/java/site/udtk/core/CoreApplication.java
@@ -1,15 +1,0 @@
-package site.udtk.core;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class CoreApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CoreApplication.class, args);
-	}
-
-}
