+++ core/src/test/java/com/underthekey/core/CoreApplicationTests.java
@@ -1,0 +1,13 @@
+package com.underthekey.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CoreApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
