--- core/src/test/java/site/udtk/core/CoreApplicationTests.java
@@ -1,13 +1,0 @@
-package site.udtk.core;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CoreApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
