+++ eureka-server/src/main/java/com/underthekey/eurekaserver/EurekaServerApplication.java
@@ -1,0 +1,17 @@
+package com.underthekey.eurekaserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication(
+	scanBasePackages = "site.udtk.globalsecurity"
+)
+@EnableEurekaServer
+public class EurekaServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EurekaServerApplication.class, args);
+	}
+
+}
