--- eureka-server/src/main/java/site/udtk/eurekaserver/EurekaServerApplication.java
@@ -1,17 +1,0 @@
-package site.udtk.eurekaserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication(
-	scanBasePackages = "site.udtk.globalsecurity"
-)
-@EnableEurekaServer
-public class EurekaServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EurekaServerApplication.class, args);
-	}
-
-}
