+++ eureka-server/src/test/java/com/underthekey/eurekaserver/EurekaServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.underthekey.eurekaserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EurekaServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
