--- eureka-server/src/test/java/site/udtk/eurekaserver/EurekaServerApplicationTests.java
@@ -1,13 +1,0 @@
-package site.udtk.eurekaserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EurekaServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
