+++ gateway-server/src/main/java/com/underthekey/gatewayserver/GatewayServerApplication.java
@@ -1,0 +1,15 @@
+package com.underthekey.gatewayserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class GatewayServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GatewayServerApplication.class, args);
+	}
+
+}
