--- gateway-server/src/main/java/site/udtk/gatewayserver/GatewayServerApplication.java
@@ -1,15 +1,0 @@
-package site.udtk.gatewayserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class GatewayServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GatewayServerApplication.class, args);
-	}
-
-}
