+++ gateway-server/src/test/java/com/underthekey/gatewayserver/GatewayServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.underthekey.gatewayserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GatewayServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
