--- gateway-server/src/test/java/site/udtk/gatewayserver/GatewayServerApplicationTests.java
@@ -1,13 +1,0 @@
-package site.udtk.gatewayserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GatewayServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
