+++ global-security/src/main/java/com/underthekey/globalsecurity/SecurityConfig.java
@@ -1,0 +1,48 @@
+package com.underthekey.globalsecurity;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+import org.springframework.security.web.SecurityFilterChain;
+
+import lombok.RequiredArgsConstructor;
+
+@Configuration
+@EnableConfigurationProperties(UserConfig.class)
+@RequiredArgsConstructor
+public class SecurityConfig {
+
+	private final UserConfig userConfig;
+
+	@Bean
+	public BCryptPasswordEncoder bcryptPasswordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+
+	@Bean
+	public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+		http
+			.csrf(AbstractHttpConfigurer::disable)
+			.authorizeHttpRequests(auth -> auth.anyRequest().authenticated())
+			.httpBasic(Customizer.withDefaults());
+		return http.build();
+	}
+
+	@Bean
+	public UserDetailsService userDetailsService() {
+		UserDetails user = User.builder()
+			.username(userConfig.getUsername())
+			.password(bcryptPasswordEncoder().encode(userConfig.getPassword()))
+			.roles(userConfig.getRole())
+			.build();
+		return new InMemoryUserDetailsManager(user);
+	}
+}
